put all non-boolean parameters in params.pp
let's just go all the way and put all of those parameters into params.
We now also set the correct group and the correct group permissions on
Debuntu. Btw, does anyone notice how on RHEL the package is libvirt, the
group is libvirt, and the service is unsurprisingly libvirtd, whereas on
Debian they are all stupid?